make DriverManager non static.move closeBrowsr to BasePage method
